+++ Userland/Libraries/LibPDF/Fonts/Type1Font.h
@@ -8,14 +8,14 @@
 
 #include <LibGfx/Font/ScaledFont.h>
 #include <LibPDF/Fonts/PDFFont.h>
-#include <LibPDF/Fonts/PS1FontProgram.h>
+#include <LibPDF/Fonts/Type1FontProgram.h>
 
 namespace PDF {
 
 class Type1Font : public PDFFont {
 public:
     struct Data : PDFFont::CommonData {
-        RefPtr<PS1FontProgram> font_program;
+        RefPtr<Type1FontProgram> font_program;
     };
 
     static PDFErrorOr<Data> parse_data(Document*, NonnullRefPtr<DictObject> font_dict, float font_size);
